+++ src/som/primitives/EqualsPrim.java
@@ -15,7 +15,7 @@ public abstract class EqualsPrim extends BinarySideEffectFreeExpressionNode {
   }
 
   @Specialization(order = 2)
-  public final boolean doInteger(final int left, final int right) {
+  public final boolean doInteger(final long left, final long right) {
     return left == right;
   }
 
@@ -45,37 +45,37 @@ public abstract class EqualsPrim extends BinarySideEffectFreeExpressionNode {
   }
 
   @Specialization(order = 100)
-  public final boolean doInteger(final int left, final double right) {
+  public final boolean doInteger(final long left, final double right) {
     return left == right;
   }
 
   @Specialization(order = 1000)
-  public final boolean doBigInteger(final BigInteger left, final int right) {
+  public final boolean doBigInteger(final BigInteger left, final long right) {
     return doBigInteger(left, BigInteger.valueOf(right));
   }
 
   @Specialization(order = 1010)
-  public final boolean doInteger(final int left, final BigInteger right) {
+  public final boolean doInteger(final long left, final BigInteger right) {
     return doBigInteger(BigInteger.valueOf(left), right);
   }
 
   @Specialization(order = 1020)
-  public final boolean doDouble(final double left, final int right) {
+  public final boolean doDouble(final double left, final long right) {
     return doDouble(left, (double) right);
   }
 
   @Specialization(order = 10000)
-  public final boolean doInteger(final int left, final String right) {
+  public final boolean doInteger(final long left, final String right) {
     return false;
   }
 
   @Specialization(order = 10010)
-  public final boolean doInteger(final int left, final SObject right) {
+  public final boolean doInteger(final long left, final SObject right) {
     return false;
   }
 
   @Specialization(order = 10100)
-  public final boolean doString(final String receiver, final int argument) {
+  public final boolean doString(final String receiver, final long argument) {
     return false;
   }
